backend/requirements.txt
Django==5.2.11
django-cors-headers==4.9.0
django-extensions==4.1
django-ninja==1.5.3
django-tenants==3.10.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pyjwt